Adds a project report
The project reports contains information on different CNN models that were tested and results on how each of them performed. The focus was on choosing between overlapping and non-overlapping pooling windows, and also on the size of Pooling windows. As a result, a model with 2x2 non-overlapping pooling windows was most accurate